refactor(web): move render helpers into translation module
