+++ packages/web/src/translation/index.ts
@@ -1,3 +1,4 @@
 export * from './effects';
 export * from './content';
 export * from './log';
+export * from './render';
